Created GameBoard class and GameBoard test
Initialized variables board and missed attacks as
a 2-D array with 10 rows and 10 columns. Initialized
variables ships and sunkShips to 0. Created getter
methods for each variable. Created a function called
placeShip that takes a ship object and coordinates and
places the ship on the board. Created a function called
receiveAttack that takes a set of coordinates and returns if the attack
was a hit or miss. Created function called allShipsSunk to return if
all ships on the board are sunk.